+++ pkg/infra/tracing/exporter.go
@@ -1,0 +1,7 @@
+package tracing
+
+import tracesdk "go.opentelemetry.io/otel/sdk/trace"
+
+type ExporterConfig interface {
+	ToExporter() (tracesdk.SpanExporter, error)
+}
